[feature] refactor templates & add mode type condition
